requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
numpy==1.20.3
phonenumbers==8.12.24
Pillow==8.2.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
